+++ CHANGELOG.md
@@ -16,6 +16,7 @@ as necessary. Empty sections will not end in the release notes.
   1. Specify the new configuration property: `nessie.version.store.persist.jdbc.datasource=mariadb`;
   2. Provide all the MariaDB (or MySQL) connection details using `quarkus.datasource.mariadb.*`
      configuration properties.
+- The Nessie GC tool is now also compatible with MariaDB and MySQL (using the MariaDB connector).
 - The Nessie Server Admin tool is now also compatible with MariaDB and MySQL (using the MariaDB
   connector).